Fixing double -> decimal

+++ db/migrate/20120927195906_add_version_to_result.rb
@@ -1,6 +1,6 @@
 class AddVersionToResult < ActiveRecord::Migration
   def change
-    add_column :results, :version, :double
+    add_column :results, :version, :decimal
 
   end
 end
